[SERVE] E2E Fix: test_standalone_3.py (#55685)
## Why are these changes needed?

Current tests are setup only to test the code when `DeploymentMode ==
EveryNode`. In this case, we have proxies on each node. When the mode is
overwritten with `HeadOnly` for any reason whatsoever, the test suite
fails.

This change enables assertions in both deployment modes.

## Related issue number

<!-- For example: "Closes #1234" -->

## Checks

- [ ] I've signed off every commit(by using the -s flag, i.e., `git
commit -s`) in this PR.
- [ ] I've run `scripts/format.sh` to lint the changes in this PR.
- [ ] I've included any doc changes needed for
https://docs.ray.io/en/master/.
- [ ] I've added any new APIs to the API Reference. For example, if I
added a
method in Tune, I've added it in `doc/source/tune/api/` under the
           corresponding `.rst` file.
- [ ] I've made sure the tests are passing. Note that there might be a
few flaky tests, see the recent failures at https://flakey-tests.ray.io/
- Testing Strategy
   - [ ] Unit tests
   - [ ] Release tests
   - [ ] This PR is not tested :(

---------
