+++ python/ray/serve/tests/test_standalone_3.py
@@ -22,6 +22,13 @@ from ray.tests.conftest import call_ray_stop_only  # noqa: F401
 from ray.util.state import list_actors
 
 
+# Some tests are not possible to run if proxy is not available on every node.
+# We skip them if proxy is not available.
+def is_proxy_on_every_node() -> bool:
+    client = _get_global_client()
+    return client._http_config.location == "EveryNode"
+
+
 @pytest.fixture
 def shutdown_ray():
     if ray.is_initialized():
@@ -286,8 +293,10 @@ def test_autoscaler_shutdown_node_http_everynode(
 
     serve.run(A.bind(), name="app_f")
 
-    # 2 proxies, 1 controller, 2 replicas.
-    wait_for_condition(lambda: len(list_actors()) == 5)
+    # If proxy is on every node, total actors are 2 proxies, 1 controller, 2 replicas.
+    # Otherwise, total actors are 1 proxy, 1 controller, 2 replicas.
+    expected_actors = 5 if is_proxy_on_every_node() else 4
+    wait_for_condition(lambda: len(list_actors()) == expected_actors)
     assert len(ray.nodes()) == 2
 
     # Stop all deployment replicas.
@@ -324,82 +333,6 @@ def test_autoscaler_shutdown_node_http_everynode(
     ray.shutdown()
 
 
-def test_drain_and_undrain_http_proxy_actors(
-    monkeypatch, shutdown_ray, call_ray_stop_only  # noqa: F811
-):
-    """Test the state transtion of the proxy actor between
-    HEALTHY, DRAINING and DRAINED
-    """
-    monkeypatch.setenv("RAY_SERVE_PROXY_MIN_DRAINING_PERIOD_S", "10")
-
-    cluster = Cluster()
-    head_node = cluster.add_node(num_cpus=0)
-    cluster.add_node(num_cpus=1)
-    cluster.add_node(num_cpus=1)
-    cluster.wait_for_nodes()
-    ray.init(address=head_node.address)
-    serve.start(http_options={"location": "EveryNode"})
-
-    @serve.deployment
-    class HelloModel:
-        def __call__(self):
-            return "hello"
-
-    serve.run(HelloModel.options(num_replicas=2).bind())
-
-    # 3 proxies, 1 controller, 2 replicas.
-    wait_for_condition(lambda: len(list_actors()) == 6)
-    assert len(ray.nodes()) == 3
-
-    client = _get_global_client()
-    serve_details = ServeInstanceDetails(
-        **ray.get(client._controller.get_serve_instance_details.remote())
-    )
-    proxy_actor_ids = {proxy.actor_id for _, proxy in serve_details.proxies.items()}
-    assert len(proxy_actor_ids) == 3
-
-    serve.run(HelloModel.options(num_replicas=1).bind())
-    # 1 proxy should be draining
-
-    def check_proxy_status(proxy_status_to_count):
-        serve_details = ServeInstanceDetails(
-            **ray.get(client._controller.get_serve_instance_details.remote())
-        )
-        proxy_status_list = [proxy.status for _, proxy in serve_details.proxies.items()]
-        print("all proxies!!!", [proxy for _, proxy in serve_details.proxies.items()])
-        current_status = {
-            status: proxy_status_list.count(status) for status in proxy_status_list
-        }
-        return current_status == proxy_status_to_count, current_status
-
-    wait_for_condition(
-        condition_predictor=check_proxy_status,
-        proxy_status_to_count={ProxyStatus.HEALTHY: 2, ProxyStatus.DRAINING: 1},
-    )
-
-    serve.run(HelloModel.options(num_replicas=2).bind())
-    # The draining proxy should become healthy.
-    wait_for_condition(
-        condition_predictor=check_proxy_status,
-        proxy_status_to_count={ProxyStatus.HEALTHY: 3},
-    )
-    serve_details = ServeInstanceDetails(
-        **ray.get(client._controller.get_serve_instance_details.remote())
-    )
-    {proxy.actor_id for _, proxy in serve_details.proxies.items()} == proxy_actor_ids
-
-    serve.run(HelloModel.options(num_replicas=1).bind())
-    # 1 proxy should be draining and eventually be drained.
-    wait_for_condition(
-        condition_predictor=check_proxy_status,
-        timeout=40,
-        proxy_status_to_count={ProxyStatus.HEALTHY: 2},
-    )
-
-    # Clean up serve.
-    serve.shutdown()
-
-
 @pytest.mark.parametrize("wait_for_controller_shutdown", (True, False))
 def test_controller_shutdown_gracefully(
     shutdown_ray, call_ray_stop_only, wait_for_controller_shutdown  # noqa: F811
@@ -426,8 +359,10 @@ def test_controller_shutdown_gracefully(
     model = HelloModel.bind()
     serve.run(target=model)
 
-    # Ensure total actors of 2 proxies, 1 controller, and 2 replicas
-    wait_for_condition(lambda: len(list_actors()) == 5)
+    # If proxy is on every node, total actors are 2 proxies, 1 controller, and 2 replicas
+    # Otherwise, total actors are 1 proxy, 1 controller, and 2 replicas
+    expected_actors = 5 if is_proxy_on_every_node() else 4
+    wait_for_condition(lambda: len(list_actors()) == expected_actors)
     assert len(ray.nodes()) == 2
 
     # Call `graceful_shutdown()` on the controller, so it will start shutdown.
@@ -485,8 +420,11 @@ def test_client_shutdown_gracefully_when_timeout(
     model = HelloModel.bind()
     serve.run(target=model)
 
-    # Ensure total actors of 2 proxies, 1 controller, and 2 replicas
-    wait_for_condition(lambda: len(list_actors()) == 5)
+    # Check expected actors based on mode
+    # If proxy is on every node, total actors are 2 proxies, 1 controller, and 2 replicas
+    # Otherwise, total actors are 1 proxy, 1 controller, and 2 replicas
+    expected_actors = 5 if is_proxy_on_every_node() else 4
+    wait_for_condition(lambda: len(list_actors()) == expected_actors)
     assert len(ray.nodes()) == 2
 
     # Ensure client times out if the controller does not shutdown within timeout.
